feat: findPrivateMapArticles 빈 메서드 생성
